fix: SimaiParser tick error when quaver is not divisible by 384
In the original code, tickStamp was calculated by `384/quaver`, whose unit is tick. However, When the quaver value was not divisible by 384, this caused loss of decimal precision, leading to incorrect timing for all subsequent notes. This error would accumulate progressively along the chart, potentially causing significant deviation in the timing of later notes.

This commit introduces a high-precision `Fraction` class and changes the rolling counter from the (int)`bar`&`tick` to (Fraction)timeInBar, to avoid decimal presicion loss.